server/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "backend of neuscheduler"
authors = [
    {name = "jacklvd",email = "vodanglong3103@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "strawberry-graphql[debug-server] (>=0.278.0,<0.279.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "openai (>=1.97.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
